Ajustes dos pacotes para organização do código
