+++ src/main/java/com/fiap/parquimetro/condutor/entity/Usuario.java
@@ -1,0 +1,39 @@
+package com.fiap.parquimetro.condutor.entity;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fiap.parquimetro.cliente.entity.enumerations.TipoUsuarioEnum;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+
+@Entity(name = "usuario")
+@JsonIgnoreProperties({ "hibernateLazyInitializer", "handler" })
+public class Usuario {
+
+   @Id
+   @GeneratedValue(strategy = GenerationType.IDENTITY)
+   private long id;
+
+   @Column(nullable = false, unique = true)
+   private String login;
+
+   @Column(nullable = false)
+   private String senha;
+
+   @Column(nullable = false)
+   private String nome;
+
+   @Enumerated(EnumType.STRING)
+   private TipoUsuarioEnum tipoUsuario;
+
+}
