+++ src/main/java/com/fiap/parquimetro/condutor/entity/enumerations/SexoEnum.java
@@ -1,0 +1,9 @@
+package com.fiap.parquimetro.condutor.entity.enumerations;
+
+public enum SexoEnum {
+    
+    MASCULINO,
+    FEMININO,
+    NA;
+    
+}
