+++ src/main/java/com/fiap/parquimetro/condutor/entity/enumerations/TipoUsuarioEnum.java
@@ -1,0 +1,8 @@
+package com.fiap.parquimetro.condutor.entity.enumerations;
+
+public enum TipoUsuarioEnum {
+
+   CLIENTE,
+   OPERADOR;
+   
+}
