+++ src/main/java/com/fiap/parquimetro/modelo/entity/Modelo.java
@@ -1,29 +1,22 @@
-package com.fiap.parquimetro.cliente.entity;
+package com.fiap.parquimetro.modelo.entity;
 
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fiap.parquimetro.marca.entity.Marca;
 
-import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
-
 @Entity
-@Table(name = "modelo")
-@JsonIgnoreProperties({ "hibernateLazyInitializer", "handler" })
 public class Modelo {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-
-    @Column(nullable = false)
     private String nome;
 
     @ManyToOne
