+++ src/main/java/com/fiap/parquimetro/tiquete/entity/Tiquete.java
@@ -1,10 +1,13 @@
-package com.fiap.parquimetro.cliente.entity;
+package com.fiap.parquimetro.tiquete.entity;
 
 import java.time.LocalDateTime;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fiap.parquimetro.cliente.entity.enumerations.StatusTiqueteEnum;
-import com.fiap.parquimetro.cliente.entity.enumerations.TipoTiqueteEnum;
+import com.fiap.parquimetro.condutor.entity.Condutor;
+import com.fiap.parquimetro.preco.entity.Preco;
+import com.fiap.parquimetro.tiquete.entity.enumerations.StatusTiqueteEnum;
+import com.fiap.parquimetro.tiquete.entity.enumerations.TipoTiqueteEnum;
+import com.fiap.parquimetro.veiculo.entity.Veiculo;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
