+++ src/main/java/com/fiap/parquimetro/tiquete/entity/enumerations/HorarioTiqueteEnum.java
@@ -1,4 +1,4 @@
-package com.fiap.parquimetro.cliente.entity.enumerations;
+package com.fiap.parquimetro.tiquete.entity.enumerations;
 
 import lombok.Getter;
