+++ src/main/java/com/fiap/parquimetro/veiculo/entity/enumerations/NacionalidadeVeiculoEnum.java
@@ -1,4 +1,4 @@
-package com.fiap.parquimetro.cliente.entity.enumerations;
+package com.fiap.parquimetro.veiculo.entity.enumerations;
 
 import lombok.Getter;
